+++ packages/react/src/index.ts
@@ -1,3 +1,9 @@
 /**
  * @3plate/graph-react - React components for @3plate/graph graph visualization
  */
+
+export { Graph } from './Graph'
+export { Playground } from './Playground'
+
+export type { GraphProps } from './Graph'
+export type { PlaygroundProps } from './Playground'