+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # @3plate/graph
 
+## 0.1.7
+
+### Patch Changes
+
+- Implement React component
+- Updated dependencies
+  - @3plate/graph-react@0.1.7
+  - @3plate/graph-core@0.1.7
+  - @3plate/graph-angular@0.1.7
+  - @3plate/graph-vue@0.1.7
+
 ## 0.1.6
 
 ### Patch Changes